+++ src/main/java/nextstep/subway/dto/LineRequest.java
@@ -7,12 +7,12 @@ public class LineRequest {
     private String color;
     private long upStationId;
     private long downStationId;
-    private long distance;
+    private int distance;
 
     public LineRequest() {
     }
 
-    public LineRequest(String name, String color, long upStationId, long downStationId, long distance) {
+    public LineRequest(String name, String color, long upStationId, long downStationId, int distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -36,7 +36,7 @@ public class LineRequest {
         return downStationId;
     }
 
-    public long getDistance() {
+    public int getDistance() {
         return distance;
     }
